:pick: Remove unneeded cast for expansion
I'll need to put the cast back in once I support a more configurable expansion factor that may take on floating-point values.

+++ src/vector.c
@@ -350,9 +350,9 @@ static bool LocalVectorExpand( struct Vector_S * self )
    {
       new_capacity = DEFAULT_INITIAL_CAPACITY;
    }
-   else if ( (uint32_t)(self->capacity * EXPANSION_FACTOR) < self->max_capacity )
+   else if ( (self->capacity * EXPANSION_FACTOR) < self->max_capacity )
    {
-      new_capacity = (uint32_t)(self->capacity * EXPANSION_FACTOR);
+      new_capacity = self->capacity * EXPANSION_FACTOR;
    }
    else
    {
